// src/app/offset/models/offsetSample.ts
export class OffsetSample {
  samplename: any;
  companyid: any;
    brandid: any;
    programid: any;
    imagepath : any;
  productcategoryid: any;
  productsubcategoryid: any;
  clientid: any;
  executiveid: any;
    specialinstruction: any;
  designerid: any;
  salesid: any;
  orderdate: any;
  estdeliverydate: any;
  approveldate: any;
  length: any;
  width: any;
  pcsyard: any;
  unit: any;
  genspcStatus: any;
  totalnoofcolor: any;
  papergsm: any;
  groundcolor: any;
  textcolor: any;
  isbarcode: boolean;
  issubcontact: number;
  isjobformultimaterial: boolean;
  submissiondate: any;
  rejectiondate: any;
  remark1: any;
    expiryStatus: boolean;
  submissioncomment: any;
  rejectioncomment: any;
  artworklocation: any;
  filename: any;
  partno : any;
  submissiondate2: any;
  submissiondate3: any;
  submissiondate4: any;
  submissioncomment2: any;
  submissioncomment3: any;
  submissioncomment4: any;
    rejectiondate2: any;
    rejectiondate3: any;
    rejectiondate4: any;
    rejectioncomment2: any;
    rejectioncomment3: any;
    rejectioncomment4: any;
    expireDate: any;
    deliverydate: any;
    completeDate: any;
    finishlength: any;
    reasonOfChange: any;
  constructor() {
    this.samplename = '';
    this.reasonOfChange = '';
    this.companyid = 0 ;
      this.imagepath = '';
      this.brandid = 0;
      this.programid = 0;
      this.clientid = 0;
      this.executiveid = 0;
      this.specialinstruction = '';
      this.designerid = 0;
    this.salesid = 0;
    this.orderdate = new Date();
    this.estdeliverydate = new Date();
    this.approveldate = '';
    this.length = 0;
    this.finishlength = 0;
    this.width = 0;
    this.unit = 0 ;
    this.genspcStatus = '';
    this.totalnoofcolor = 0;
   this.papergsm = 0;
    this.groundcolor = '';
    this.textcolor = '';
    this.isbarcode = false;
    this.isjobformultimaterial = false;
    this.issubcontact = 0;
    this.expiryStatus = false;
    this.submissiondate = '';
    this.expireDate = '';
    this.deliverydate = '';
    this.completeDate = '';
    this.rejectiondate = '';
    this.remark1 = '';
    this.submissioncomment =  '';
    this.rejectioncomment =  '';
    this.artworklocation = '';
    this.filename = '';
    this.submissiondate2 = '';
    this.submissiondate3 = '';
    this.submissiondate4 = '';
    this.submissioncomment2 = '';
    this.submissioncomment3 = '';
    this.submissioncomment4 = '';
    this.partno = 1;
      this.rejectioncomment2 =  '';
      this.rejectioncomment3 =  '';
      this.rejectioncomment4 =  '';
      this.rejectiondate2 =  '';
      this.rejectiondate3 = '';
      this.rejectiondate4 = '';
  }
}